fix(find-make): suppress printing gmake path on every invocation
`locate` (and `command -v`) add noise to the logs for programs invoking
shystemctl through printing the path of the program if it's found during
every invocation of the command.

It looks like the usage of `locate` is meant to check if `gmake` exists
and exit with the corresponding exit code. But `locate` successfully
exits whether the command is found or not. So I switched to using
`command -v` as a POSIX compatible way of doing the check, along with
suppressing the output.